Add files via upload
The organizer of a senior citizens’ club arranges outings for the members. For each of these outings, a
coach is hired, meals at a restaurant are reserved and tickets for the theatre are booked. A program is
required to work out the costs and provide a printed list showing everyone on the outing.
Write and test a program for the club organizer.
Your program must include appropriate prompts for the entry of data.
Error messages and other outputs need to be set out clearly.
All variables, constants and other identifiers must have meaningful names.
You will need to complete these three tasks. Each task must be fully tested.
TASK 1 – Work out the total cost of the outing.
The organizer finds out how many senior citizens would be interested in the outing. The program for
TASK 1 works out the cost from this information.

The minimum number of senior citizens needed for the outing to go ahead is 10; there cannot be more
than 36 senior citizens on the outing. A minimum of two carers must go on the outing, with an additional
carer needed if more than 24 senior citizens go on the outing. Carers do not have to pay anything for
the outing. Work out the total cost and the cost per person for the senior citizens.
TASK 2 – Record who is going on the outing and how much has been paid.
Using your results from TASK 1, record the names of the people on the outing and the amount they
have paid; include the carers on the outing. If there are spare places on the coach then extra people
can be added; they are charged the same price as the other senior citizens. Calculate the total amount
of money collected. Print out a list of the people on the outing.
TASK 3 – Identify the break-even point or profit that will be made on the outing.
Show whether the outing has made a profit or has broken even using the estimated cost from TASK 1
and the money collected from TASK 2.